fix: AgeDropDown, InterestDropdown 컴포넌트로 통합 및 불필요한 파일 제거
